macos: use xcodeproj
